requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Templates
jinja2==3.1.3

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2

# PDF Generation
reportlab==4.1.0

# Date handling
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
aiofiles==23.2.1
